requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
executing==2.1.0
fastapi==0.114.0
fastapi-cli==0.0.5
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-community==0.3.1
langchain-core==0.3.8
langchain-text-splitters==0.3.0
langgraph==0.2.34
langgraph-checkpoint==2.0.0
langgraph-checkpoint-postgres==2.0.0
langsmith==0.1.129
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-pool==3.2.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.5
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
yarl==1.13.1
